api/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
python-dotenv==1.0.1
redis==5.0.8
celery==5.4.0